# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte ETV",
        "description": "REST API. Projecte ETV. DAW Client i servidor.",
        "contact": {
            "name": "Webmaster.",
            "email": "info@webmaster.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://www.dawpaucasesnoves.com/etvServidor/public"
        }
    ],
    "paths": {
        "/api/arquitectes": {
            "get": {
                "tags": [
                    "Arquitectes"
                ],
                "summary": "Mostrar els arquitectes paginats",
                "description": "Display a listing of the resource.",
                "operationId": "f14ff3a8e25dacccb21effaaec33fa44",
                "responses": {
                    "200": {
                        "description": "Mostrar tots els arquitectes de forma paginada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Arquitectes"
                ],
                "summary": "Crea un arquitecte",
                "description": "Crea un nou arquitecte.",
                "operationId": "67fa059629e7ffc52bdd6353d49cf082",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Lluís Domènech i Montaner"
                                    },
                                    "data_naix": {
                                        "type": "string",
                                        "format": "yyyy-mm-dd",
                                        "example": "1850-12-21"
                                    },
                                    "descripcio": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Lluís Domènech i Montaner va ser un destacat arquitecte espanyol del modernisme català, nascut el 1850 i mort el 1923.\n     *                                                                                               És conegut per obres emblemàtiques com l'Hospital de Sant Pau i el Palau de la Música Catalana a\n     *                                                                                                   Barcelona, que reflecteixen la seva contribució significativa al desenvolupament arquitectònic de la ciutat durant el segle XIX i principis del XX."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        },
                                        "missatge": {
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string"
                                        },
                                        "codi": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/arquitectes_tots": {
            "get": {
                "tags": [
                    "Arquitectes"
                ],
                "summary": "Mostrar els arquitectes",
                "description": "Display a listing of the resource.",
                "operationId": "3971c83dd61616cd6a23cd0ed3443e20",
                "responses": {
                    "200": {
                        "description": "Mostrar tots els arquitectes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/arquitectes/find/{cerca}": {
            "get": {
                "tags": [
                    "Arquitectes"
                ],
                "summary": "Mostrar els arquitectes que contenguin el filtre de manera paginada",
                "operationId": "d1a96878f92e8e0b0ea2a8c772e86cea",
                "parameters": [
                    {
                        "name": "cerca",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}